Combine the completion.json and functions.json editor support files into a single editor.json file
